Use multithreading to speed up the search
Multiprocessing would achieve higher speedups as it would maximize the
concurrent operations done by the available cores without being slowed
down by the GIL. However, because the size of the ANN index objects is
in the order of several GBs this prevents memory duplication.
Instead, the slower, shared-memory, multithreading is used.