Remove NMSLIB support
Focus on providing optimal support for Annoy instead.

ANN indexing through NMSLIB incurs a massive overhead as the item
database has to be reinitialized every time. If in the future NMSLIB
supports memmapping (see searchivarius/nmslib#133) we can consider to
add NMSLIB again based in this previous code.

Also switch to Python 3.6 for the Docker container now that we aren't
restricted to Python 2 by NMSLIB anymore.